requirements.txt
moviepy==1.0.3
gtts==2.5.3
requests==2.32.3
psutil==6.0.0
pillow==10.4.0